requirements.txt
python-dotenv
requests
schedule
pywhatkit
flask